% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_multidesign.R
\name{bind_multidesign}
\alias{bind_multidesign}
\title{Combine multiple multidesign objects}
\usage{
bind_multidesign(..., .id = NULL)
}
\arguments{
\item{...}{multidesign objects to combine}

\item{.id}{optional name of an identifier column added to the design
to indicate the origin multidesign}
}
\value{
A new multidesign object containing the concatenated data
}
\description{
This function row-binds the observation matrices and design data frames
of several multidesign objects. All input multidesigns must share the
same column design. Optionally, an identifier column can be added to
track the source of each observation. Explicit cell masks are row-bound; if
only some inputs are masked, unmasked inputs are treated as all observed.
}
\examples{
X1 <- matrix(rnorm(10*5), 10, 5)
X2 <- matrix(rnorm(10*5), 10, 5)
md1 <- multidesign(X1, data.frame(cond = rep("A", 10)))
md2 <- multidesign(X2, data.frame(cond = rep("B", 10)))
combined <- bind_multidesign(md1, md2)
combined_id <- bind_multidesign(md1, md2, .id = "source")

}
