% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hyperdesign.R
\name{block_indices}
\alias{block_indices}
\alias{block_indices.hyperdesign}
\title{Get Block Indices from a Multiblock Object}
\usage{
block_indices(x, ...)

\method{block_indices}{hyperdesign}(x, i, byrow = FALSE, ...)
}
\arguments{
\item{x}{The object to get block indices from}

\item{...}{Additional arguments passed to methods}

\item{i}{the block number}

\item{byrow}{if true, return row-oriented indices}
}
\value{
An integer vector or list of integer vectors with indices for the
  requested block(s)
}
\description{
This is a re-export of the block_indices function from the multivarious package.
See \code{multivarious::block_indices} for full documentation.
}
\details{
Get Block Indices
}
\examples{
mb <- multiblock(list(matrix(1:10, 5, 2), matrix(11:20, 5, 2)))
block_indices(mb, 1)
block_indices(mb, 2)

}
