% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdesign.R
\name{column_design.hyperdesign}
\alias{column_design.hyperdesign}
\title{Get Column Design Information for Hyperdesign}
\usage{
\method{column_design}{hyperdesign}(x, block, ...)
}
\arguments{
\item{x}{A hyperdesign object}

\item{block}{Optional block index or name to get design for specific block}

\item{...}{Additional arguments passed to methods}
}
\value{
A list of column design information for each bloc
}
\description{
Extract column design information from a hyperdesign object.
}
