% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesign.R
\name{column_design.multidesign}
\alias{column_design.multidesign}
\title{Extract the column design from a multidesign object}
\usage{
\method{column_design}{multidesign}(x, ...)
}
\arguments{
\item{x}{A multidesign object}

\item{...}{Additional arguments passed to methods}
}
\value{
The column design data frame
}
\description{
Extract the column design from a multidesign object
}
