% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence.R
\name{column_space}
\alias{column_space}
\title{Get the Column-Space Contract of a Hyperdesign}
\usage{
column_space(x)
}
\arguments{
\item{x}{A hyperdesign object.}
}
\value{
`"common"`, `"block"`, or `NULL` when unspecified.
}
\description{
Get the Column-Space Contract of a Hyperdesign
}
