% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesign.R
\name{design.multidesign}
\alias{design.multidesign}
\title{Extract Design Information from a Multidesign Object}
\usage{
\method{design}{multidesign}(x, ...)
}
\arguments{
\item{x}{A multidesign object}

\item{...}{Additional arguments (not used)}
}
\value{
The design data frame stored in the multidesign object
}
\description{
Returns the design data frame component of a multidesign object.
}
\seealso{
Other multidesign functions: 
\code{\link{multidesign}()},
\code{\link{reduce.multidesign}()},
\code{\link{split.multidesign}()},
\code{\link{split_indices.multidesign}()},
\code{\link{subset.multidesign}()},
\code{\link{summarize_by.multidesign}()},
\code{\link{xdata.multidesign}()}
}
\concept{multidesign functions}
