% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correspondence.R
\name{entity_id}
\alias{entity_id}
\title{Get the Entity ID Column of a Hyperdesign}
\usage{
entity_id(x)
}
\arguments{
\item{x}{A hyperdesign object.}
}
\value{
The design-column name, or `NULL` when unset or positional.
}
\description{
Get the Entity ID Column of a Hyperdesign
}
