% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{is_rstacked.multiblock_list}
\alias{is_rstacked.multiblock_list}
\title{Test if Multiblock Object is Row-Stacked}
\usage{
\method{is_rstacked}{multiblock_list}(x)
}
\arguments{
\item{x}{A multiblock object}
}
\value{
Logical indicating if object is row-stacked
}
\description{
Checks if a multiblock object is row-stacked (matrices share column dimension).
}
\seealso{
[is_cstacked()]

Other multiblock functions: 
\code{\link{block_indices.multiblock_list}()},
\code{\link{is_cstacked.multiblock_list}()},
\code{\link{multiblock.list}()},
\code{\link{t.multiblock_list}()}
}
\concept{multiblock functions}
