% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{is_rstacked}
\alias{is_rstacked}
\title{Test if Multiblock Object is Row Stacked}
\usage{
is_rstacked(x)
}
\arguments{
\item{x}{The multiblock object to test}
}
\value{
Logical value: TRUE if the object is row-stacked, FALSE otherwise
}
\description{
Checks if a multiblock object is row-stacked, meaning that all component matrices
share the same number of columns. In row-stacked multiblock objects, blocks are arranged
one above another, with each block potentially having a different number of rows.
}
\examples{
# Create row-stacked multiblock (matrices share column dimension)
Y1 <- matrix(rnorm(10*5), 10, 5)
Y2 <- matrix(rnorm(15*5), 15, 5)
mb <- multiblock(list(Y1, Y2))

# Test stacking orientation
is_rstacked(mb)  # Returns TRUE
is_cstacked(mb)  # Returns FALSE

}
\seealso{
\code{\link{is_cstacked}} for testing if a multiblock object is column-stacked,
  \code{\link{multiblock}} for creating multiblock objects
}
