% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{multiblock}
\alias{multiblock}
\title{create an multiblock object}
\usage{
multiblock(x, design, ...)
}
\arguments{
\item{x}{the matrix data}

\item{design}{design matrix for the shared dimension}

\item{...}{extra args}
}
\description{
construct a new multiblock matrix consisting of a set of stacked submatrices sharing a row- or column dimension.
}
\examples{

X <- replicate(3, matrix(rnorm(20*100), 20, 100), simplify=FALSE)
mb <- multiblock(X)
}
