% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{observation.deflist}
\alias{observation.deflist}
\title{Create Observation from Deflist}
\usage{
\method{observation}{deflist}(x, i)
}
\arguments{
\item{x}{A deflist object}

\item{i}{Index of the observation to create}
}
\value{
An observation object (function) that returns the data when called
}
\description{
Creates a lazy-evaluated observation from a deflist object.
}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.list}()},
\code{\link{observation.matrix}()},
\code{\link{observation.vector}()},
\code{\link{split.multiframe}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
