% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{observation.list}
\alias{observation.list}
\title{Create Observation from List}
\usage{
\method{observation}{list}(x, i)
}
\arguments{
\item{x}{A list}

\item{i}{Index of the list element to create as an observation}
}
\value{
An observation object (function) that returns the list element when called
}
\description{
Creates a lazy-evaluated observation from a list element.
}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.matrix}()},
\code{\link{observation.vector}()},
\code{\link{split.multiframe}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
