% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{observation.matrix}
\alias{observation.matrix}
\title{Create Observation from Matrix}
\usage{
\method{observation}{matrix}(x, i)
}
\arguments{
\item{x}{A matrix}

\item{i}{Row index of the observation to create}
}
\value{
An observation object (function) that returns the matrix row when called
}
\description{
Creates a lazy-evaluated observation from a row of a matrix.
}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.list}()},
\code{\link{observation.vector}()},
\code{\link{split.multiframe}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
