% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{observation.vector}
\alias{observation.vector}
\title{Create Observation from Vector}
\usage{
\method{observation}{vector}(x, i)
}
\arguments{
\item{x}{A vector}

\item{i}{Index (ignored for vectors, as they represent a single observation)}
}
\value{
An observation object (function) that returns the vector when called
}
\description{
Creates a lazy-evaluated observation from a vector.
}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.list}()},
\code{\link{observation.matrix}()},
\code{\link{split.multiframe}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
