% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{print.aligned_hyperdesign}
\alias{print.aligned_hyperdesign}
\title{Print an Aligned Hyperdesign}
\usage{
\method{print}{aligned_hyperdesign}(x, ...)
}
\arguments{
\item{x}{An aligned_hyperdesign object.}

\item{...}{Additional arguments (unused).}
}
\value{
`x`, invisibly.
}
\description{
Print an Aligned Hyperdesign
}
