% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hyperdesign.R, R/multidesign.R
\name{select_variables}
\alias{select_variables}
\alias{select_variables.hyperdesign}
\alias{select_variables.multidesign}
\title{Select Variables Based on Column Design}
\usage{
select_variables(x, ...)

\method{select_variables}{hyperdesign}(x, ...)

\method{select_variables}{multidesign}(x, ...)
}
\arguments{
\item{x}{The object to select variables from (multidesign, hyperdesign, etc.)}

\item{...}{Filter expressions evaluated against the column design (using dplyr::filter semantics)}
}
\value{
An object of the same class with a subset of columns/variables
}
\description{
Subsets the columns (variables) of an object based on conditions applied to the
column design metadata. This allows filtering variables by their properties
(e.g., region, type, hemisphere) using dplyr-style filter expressions.
}
\examples{
X <- matrix(rnorm(20*10), 20, 10)
Y <- data.frame(condition = rep(c("A", "B"), each=10))
col_info <- data.frame(
  region = rep(c("frontal", "parietal"), 5),
  hemisphere = rep(c("left", "right"), each=5)
)
mds <- multidesign(X, Y, col_info)

# Select only frontal variables
mds_frontal <- select_variables(mds, region == "frontal")

}
\seealso{
\code{\link{column_design}} for extracting column metadata,
  \code{\link{multidesign}} for creating multidesign objects
}
