% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesign.R
\name{split.multidesign}
\alias{split.multidesign}
\title{Split a Multidesign Object by Design Variables}
\usage{
\method{split}{multidesign}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{A multidesign object}

\item{f}{Unquoted name of the first variable to split by}

\item{drop}{Logical; unused, present for compatibility with the generic}

\item{...}{Additional unquoted names of variables to split by}
}
\value{
A list of multidesign objects, one for each combination of splitting variables
}
\description{
Splits a multidesign object into multiple objects based on combinations of one or more
design variables.
}
\examples{
X <- matrix(rnorm(100*20), 100, 20)
Y <- tibble::tibble(
  condition = rep(c("A", "B"), each=50),
  block = rep(1:2, times=50)
)
mds <- multidesign(X, Y)

# Split by condition
split_by_cond <- split(mds, condition)

# Split by condition and block
split_by_both <- split(mds, condition, block)

}
\seealso{
\code{\link{subset.multidesign}}

Other multidesign functions: 
\code{\link{multidesign}()},
\code{\link{reduce.multidesign}()},
\code{\link{split_indices.multidesign}()},
\code{\link{subset.multidesign}()},
\code{\link{summarize_by.multidesign}()}
}
\concept{multidesign functions}
