% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{split.multiframe}
\alias{split.multiframe}
\title{Split a Multiframe by Design Variables}
\usage{
\method{split}{multiframe}(x, ..., collapse = FALSE)
}
\arguments{
\item{x}{A multiframe object}

\item{...}{Unquoted names of variables to split by}

\item{collapse}{Logical; if TRUE, combines observations in each group into a matrix}
}
\value{
A nested tibble containing the split data
}
\description{
Splits a multiframe into groups based on one or more design variables.
Optionally collapses the observations within each group into a single matrix.
}
\examples{
x <- matrix(1:40, 10, 4)
y <- data.frame(
  condition = rep(c("A", "B"), each=5),
  subject = rep(1:5, times=2)
)
mf <- multiframe(x, y)

# Split by condition
split_by_cond <- split(mf, condition)

# Split by condition and subject, collapsing observations
split_by_both <- split(mf, condition, subject, collapse=TRUE)

}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.list}()},
\code{\link{observation.matrix}()},
\code{\link{observation.vector}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
