% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{[.observation_set}
\alias{[.observation_set}
\title{Extract Multiple Observations from Observation Set}
\usage{
\method{[}{observation_set}(x, i)
}
\arguments{
\item{x}{An observation_set object}

\item{i}{Indices of observations to extract}
}
\value{
List of evaluated observations
}
\description{
Extracts multiple observations from an observation_set using single bracket indexing.
All selected observations are automatically evaluated.
}
\seealso{
Other multiframe functions: 
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.list}()},
\code{\link{observation.matrix}()},
\code{\link{observation.vector}()},
\code{\link{split.multiframe}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
