% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{summarize_by.multiframe}
\alias{summarize_by.multiframe}
\title{Summarize a Multiframe by Design Variables}
\usage{
\method{summarize_by}{multiframe}(x, ..., sfun = colMeans, extract_data = FALSE)
}
\arguments{
\item{x}{A multiframe object}

\item{...}{Unquoted names of variables to group by}

\item{sfun}{Summary function to apply (default is colMeans)}

\item{extract_data}{Logical; if TRUE, returns just the summarized data}
}
\value{
A tibble containing summarized data for each group
}
\description{
Computes summaries of observations grouped by design variables.
}
\examples{
x <- matrix(1:40, 10, 4)
y <- data.frame(
  condition = rep(c("A", "B"), each=5),
  subject = rep(1:5, times=2)
)
mf <- multiframe(x, y)

# Get means by condition
means_by_cond <- summarize_by(mf, condition)

}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{[[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.list}()},
\code{\link{observation.matrix}()},
\code{\link{observation.vector}()},
\code{\link{split.multiframe}()}
}
\concept{multiframe functions}
