% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiblock.R
\name{t.multiblock_list}
\alias{t.multiblock_list}
\title{Transpose a Multiblock Object}
\usage{
\method{t}{multiblock_list}(x)
}
\arguments{
\item{x}{A multiblock object}
}
\value{
A new multiblock object with transposed matrices
}
\description{
Transposes each matrix in a multiblock object and returns a new multiblock object.
The stacking orientation will be switched (row-stacked becomes column-stacked and vice versa).
}
\examples{
mb <- multiblock(list(matrix(1:10, 5, 2), matrix(11:20, 5, 2)))
mb_t <- t(mb)
is_cstacked(mb_t)
is_rstacked(mb_t)
}
\seealso{
Other multiblock functions: 
\code{\link{block_indices.multiblock_list}()},
\code{\link{is_cstacked.multiblock_list}()},
\code{\link{is_rstacked.multiblock_list}()},
\code{\link{multiblock.list}()}
}
\concept{multiblock functions}
