% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdesign.R
\name{xdata.hyperdesign}
\alias{xdata.hyperdesign}
\title{Extract Data Matrix from Hyperdesign}
\usage{
\method{xdata}{hyperdesign}(x, block, ...)
}
\arguments{
\item{x}{A hyperdesign object}

\item{block}{Optional numeric index specifying which block's data to return}

\item{...}{Additional arguments (not used)}
}
\value{
If block is specified, returns data matrix for that block; otherwise returns
        a list of data matrices for all blocks
}
\description{
Get the data matrix component from a hyperdesign object, either for all blocks
or for a specific block.
}
\examples{
# Create example hyperdesign
d1 <- multidesign(matrix(rnorm(10*20), 10, 20),
                  data.frame(y=1:10, condition=rep(c("A","B"), 5)))
d2 <- multidesign(matrix(rnorm(10*20), 10, 20),
                  data.frame(y=1:10, condition=rep(c("A","B"), 5)))
hd <- hyperdesign(list(d1, d2))

# Get data from all blocks
all_data <- xdata(hd)

# Get data from block 1
block1_data <- xdata(hd, block=1)
}
\seealso{
[design.hyperdesign()], [column_design.hyperdesign()]

Other hyperdesign functions: 
\code{\link{design.hyperdesign}()},
\code{\link{df_to_hyperdesign}()},
\code{\link{hyperdesign}()},
\code{\link{init_transform.hyperdesign}()},
\code{\link{subset.hyperdesign}()}
}
\concept{hyperdesign functions}
