% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesign.R
\name{xdata.multidesign}
\alias{xdata.multidesign}
\title{Extract Data Matrix from a Multidesign Object}
\usage{
\method{xdata}{multidesign}(x, ...)
}
\arguments{
\item{x}{A multidesign object}

\item{...}{Additional arguments (not used)}
}
\value{
The data matrix stored in the multidesign object
}
\description{
Returns the data matrix component of a multidesign object.
}
\seealso{
Other multidesign functions: 
\code{\link{design.multidesign}()},
\code{\link{multidesign}()},
\code{\link{reduce.multidesign}()},
\code{\link{split.multidesign}()},
\code{\link{split_indices.multidesign}()},
\code{\link{subset.multidesign}()},
\code{\link{summarize_by.multidesign}()}
}
\concept{multidesign functions}
