% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_multidesign.R
\name{bind_multidesign}
\alias{bind_multidesign}
\title{Combine multiple multidesign objects}
\usage{
bind_multidesign(..., .id = NULL)
}
\arguments{
\item{...}{multidesign objects to combine}

\item{.id}{optional name of an identifier column added to the design
to indicate the origin multidesign}
}
\value{
A new multidesign object containing the concatenated data
}
\description{
Row-binds the observation matrices and design data frames of several
multidesign objects. All inputs must share the same column design.
}
