% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hyperdesign.R
\name{block_indices}
\alias{block_indices}
\alias{block_indices.hyperdesign}
\title{Get Block Indices}
\usage{
block_indices(x, ...)

block_indices(x, ...)

\method{block_indices}{hyperdesign}(x, i, byrow = FALSE, ...)
}
\arguments{
\item{x}{The object to get block indices from}

\item{...}{Additional arguments passed to methods}

\item{i}{the block number}

\item{byrow}{if true, return row-oriented indices}
}
\value{
A matrix with start and end indices for each block
}
\description{
Retrieve the indices that define the boundaries of blocks in a multiblock or hyperdesign object.

This is a re-export of the block_indices function from the multivarious package.
See \code{multivarious::block_indices} for full documentation.
}
\details{
Get Block Indices
}
