% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hyperdesign.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{block_indices}
\alias{block_indices.hyperdesign}
\title{Objects exported from other packages}
\usage{
\method{block_indices}{hyperdesign}(x, i, byrow = FALSE)
}
\arguments{
\item{i}{the block number}

\item{byrow}{if true, return row-oriented indices}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{multivarious}{\code{\link[multivarious]{block_indices}}}
}}

