% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdesign.R
\name{df_to_hyperdesign}
\alias{df_to_hyperdesign}
\title{Function to create a hyperdesign from a data.frame (or tibble)}
\usage{
df_to_hyperdesign(data, design_vars, x_vars, split_var)
}
\arguments{
\item{data}{A data.frame.}

\item{design_vars}{a character vector of design variables to extract}

\item{x_vars}{a character vector of X design variables to extract}

\item{split_var}{the splitting variable.}
}
\value{
A hyperdesign object.
}
\description{
This function takes a data.frame, the names of the design variables, the names of the X variables,
and the name of the splitting variable, and returns a hyperdesign object.
}
\examples{
# Create a sample tibble
sample_tibble <- tibble(
  felab = rep(1:2, each = 3),
  attention = rep(c("DA", "FA"), times = 3),
  basis = rep(c("basis01", "basis02", "basis03"), times = 2),
  subject = rep(1001:1002, each = 3),
  `1` = rnorm(6),
  `2` = rnorm(6),
  `3` = rnorm(6)
)

# Apply the function
hd <- df_to_hyperdesign(sample_tibble, c("felab", "attention", "basis"), setdiff(colnames(sample_tibble), c("felab", "attention", "basis", "subject")), "subject")
}
