% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{fold_over}
\alias{fold_over}
\title{generate cross-validation folds}
\usage{
fold_over(x, ...)
}
\arguments{
\item{x}{the dataset to fold over}

\item{...}{extra args}
}
\description{
create a set of cross-validation objects over a blocking variable (`stratum`)
}
