% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hyperdesign.R
\name{fold_over}
\alias{fold_over}
\alias{fold_over.hyperdesign}
\title{generate cross-validation folds}
\usage{
fold_over(x, ...)

\method{fold_over}{hyperdesign}(x, ...)
}
\arguments{
\item{x}{the dataset to fold over}

\item{...}{extra args}
}
\description{
create a set of cross-validation objects over a blocking variable (`stratum`)
}
\examples{
d1 <- multidesign(matrix(rnorm(10*20), 10, 20), data.frame(y=1:10, subject=1, run=rep(1:5, 2)))
d2 <- multidesign(matrix(rnorm(10*20), 10, 20), data.frame(y=1:10, subject=2, run=rep(1:5, 2)))
d3 <- multidesign(matrix(rnorm(10*20), 10, 20), data.frame(y=1:10, subject=3, run=rep(1:5, 2)))

hd <- hyperdesign(list(d1,d2,d3))
folds <- fold_over(hd, run)
}
