% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/hyperdesign.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{init_transform}
\alias{init_transform.hyperdesign}
\title{Objects exported from other packages}
\usage{
\method{init_transform}{hyperdesign}(x, preproc)
}
\value{
a transformed `hyperdesign` object, with pre-processors set as attribute named `preproc`
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{multivarious}{\code{\link[multivarious]{init_transform}}}
}}

