% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{is_cstacked}
\alias{is_cstacked}
\title{test whether `multiblock` object is column stacked}
\usage{
is_cstacked(x)
}
\arguments{
\item{x}{the object}
}
\description{
test whether `multiblock` object is column stacked
}
