% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{is_rstacked}
\alias{is_rstacked}
\title{test whether `multiblock` object is row stacked}
\usage{
is_rstacked(x)
}
\arguments{
\item{x}{the object}
}
\description{
test whether `multiblock` object is row stacked
}
