% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_spec.R
\name{model_matrix}
\alias{model_matrix}
\alias{term_data}
\alias{grouping_data}
\alias{component_data}
\alias{grouping_term_data}
\alias{random_effect_data}
\alias{design_blueprint}
\alias{design_rows}
\title{Access compiled design products}
\usage{
model_matrix(x)

term_data(x)

grouping_data(x)

component_data(x)

grouping_term_data(x)

random_effect_data(x)

design_blueprint(x)

design_rows(x)
}
\arguments{
\item{x}{A `compiled_design`.}
}
\value{
The requested matrix or metadata table.
}
\description{
Access compiled design products
}
