% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{multiblock}
\alias{multiblock}
\title{create an multiblock object}
\usage{
multiblock(x, ...)
}
\arguments{
\item{x}{the matrix data}

\item{...}{extra args}

\item{design}{design matrix for the shared dimension}
}
\description{
construct a new multiblock matrix consisting of a set of stacked submatrices sharing a row- or column dimension.
}
\examples{
X <- lapply(c(10,20,30), function(i) matrix(rnorm(i*100), i, 100))
mb <- multiblock(X)
}
