% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidesign.R
\name{reduce.multidesign}
\alias{reduce.multidesign}
\title{Reduce Dimensionality of a Multidesign Object}
\usage{
reduce.multidesign(
  x,
  nc = 2,
  ...,
  rfun = function(x) multivarious::pca(x$x, ncomp = nc, ...)
)
}
\arguments{
\item{x}{A multidesign object}

\item{nc}{Number of components to retain in the reduction}

\item{...}{Additional arguments passed to rfun}

\item{rfun}{Function to perform dimensionality reduction, must return a projector object}
}
\value{
A reduced_multidesign object containing:
  \item{x}{The reduced data matrix}
  \item{design}{Original design information}
  \item{projector}{The projection object used for reduction}
}
\description{
Performs dimensionality reduction on the data matrix of a multidesign object while preserving
the design structure. By default uses PCA, but supports any reduction method that returns
a projector object.
}
\examples{
X <- matrix(rnorm(100*20), 100, 20)
Y <- tibble(condition = rep(letters[1:4], each=25))
mds <- multidesign(X, Y)

# Reduce to 5 components using PCA
reduced_mds <- reduce(mds, nc=5)

}
\seealso{
\code{\link{multidesign}}

Other multidesign functions: 
\code{\link{design.multidesign}()},
\code{\link{multidesign}()},
\code{\link{split.multidesign}()},
\code{\link{split_indices.multidesign}()},
\code{\link{subset.multidesign}()},
\code{\link{summarize_by.multidesign}()},
\code{\link{xdata.multidesign}()}
}
\concept{multidesign functions}
