% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{split_indices}
\alias{split_indices}
\title{split indices}
\usage{
split_indices(x, ...)
}
\arguments{
\item{x}{the object to split}

\item{...}{the splitting variables}
}
\description{
extract the row indices of a table, split by one or more variables
}
