% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiframe.R
\name{[[.observation_set}
\alias{[[.observation_set}
\title{Extract Single Observation from Observation Set}
\usage{
\method{[[}{observation_set}(x, i)
}
\arguments{
\item{x}{An observation_set object}

\item{i}{Index of the observation to extract}
}
\value{
The evaluated observation
}
\description{
Extracts a single observation from an observation_set using double bracket indexing.
The result is automatically evaluated.
}
\seealso{
Other multiframe functions: 
\code{\link{[.observation_set}()},
\code{\link{multiframe.list}()},
\code{\link{multiframe.matrix}()},
\code{\link{obs_group}()},
\code{\link{observation.deflist}()},
\code{\link{observation.list}()},
\code{\link{observation.matrix}()},
\code{\link{observation.vector}()},
\code{\link{split.multiframe}()},
\code{\link{summarize_by.multiframe}()}
}
\concept{multiframe functions}
